pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spending-coach"
version = "2.0.0"
description = "LangGraph-based financial analysis agent for Amazon order history"
requires-python = ">=3.9"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.15",
    "langchain-core>=0.3.0",
    "anthropic>=0.52.0",
    "python-dotenv>=1.1.0",
]

[project.optional-dependencies]
forecast-prophet = [
    "prophet>=1.1.5",
    "pandas>=2.0.0",
]
forecast-arima = [
    "statsmodels>=0.14.0",
    "pandas>=2.0.0",
]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
spending-coach = "main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["config*", "graph*"]

# Standalone modules at project root that the CLI entry point depends on
[tool.setuptools]
py-modules = ["main", "logging_config"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
